Обновление типа первичного ключа в JPA-репозиториях
Тип первичных ключей в БД и моделях был изменен, теперь вношу изменения в JPA-репозитории

+++ reference/src/main/java/com/ebito/reference/repository/ContractRepository.java
@@ -3,5 +3,5 @@ package com.ebito.reference.repository;
 import com.ebito.reference.model.Contract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 }
